\name{extract.image}
\alias{extract.image}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Extract image data from adimpro object}
\description{
  Extract image data from adimpro object
}
\usage{
extract.image(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{ adimpro object }
}
\value{
  array or matrix of image data
}
\author{ Karsten Tabelow \email{tabelow@wias-berlin.de} and Joerg Polzehl \email{polzehl@wias-berlin.de}}
\seealso{  \code{\link{make.image}} }
\examples{
\dontrun{demo(io)}
\dontrun{demo(manipulate)}
}
\keyword{ manip }% at least one, from doc/KEYWORDS
